lithopy/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
easygui==0.98.2
fonttools==4.28.2
frozenlist==1.2.0
future==0.18.2
google-search-results==2.4.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.0
multidict==5.2.0
numpy==1.21.4
numpy-stl==2.16.3
opencv-python==4.5.4.60
packaging==21.3
pathlib2==2.3.6
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-utils==2.5.6
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
vtk==9.1.0
vtkplotlib==1.4.1
wslink==1.1.1
yarl==1.7.2
